Adjust e2e test suite structure
1. Move BeforeSuite logic to main file (e2e_test).
2. Properly Skip() quota tests (file was just ignored before).
